+++ proto/example/message/v1/other.ddb.go
@@ -14,14 +14,14 @@ import (
 func (x *OtherKitchen) MarshalDynamoItem() (m map[string]types.AttributeValue, err error) {
 	m = make(map[string]types.AttributeValue)
 	if x.AnotherKitchen != nil {
-		m16, err := ddb.MarshalMessage(x.GetAnotherKitchen(), ddb.Embed(v1.Encoding_ENCODING_UNSPECIFIED))
+		m16, err := ddb.MarshalMessage(x.GetAnotherKitchen(), ddb.Embed(v1.Encoding_ENCODING_DYNAMO))
 		if err != nil {
 			return nil, fmt.Errorf("failed to marshal field 'AnotherKitchen': %w", err)
 		}
 		m["16"] = m16
 	}
 	if x.OtherTimer != nil {
-		m17, err := ddb.MarshalMessage(x.GetOtherTimer(), ddb.Embed(v1.Encoding_ENCODING_UNSPECIFIED))
+		m17, err := ddb.MarshalMessage(x.GetOtherTimer(), ddb.Embed(v1.Encoding_ENCODING_DYNAMO))
 		if err != nil {
 			return nil, fmt.Errorf("failed to marshal field 'OtherTimer': %w", err)
 		}
@@ -34,14 +34,14 @@ func (x *OtherKitchen) MarshalDynamoItem() (m map[string]types.AttributeValue, e
 func (x *OtherKitchen) UnmarshalDynamoItem(m map[string]types.AttributeValue) (err error) {
 	if m["16"] != nil {
 		x.AnotherKitchen = new(Kitchen)
-		err = ddb.UnmarshalMessage(m["16"], x.AnotherKitchen, ddb.Embed(v1.Encoding_ENCODING_UNSPECIFIED))
+		err = ddb.UnmarshalMessage(m["16"], x.AnotherKitchen, ddb.Embed(v1.Encoding_ENCODING_DYNAMO))
 		if err != nil {
 			return fmt.Errorf("failed to unmarshal field 'AnotherKitchen': %w", err)
 		}
 	}
 	if m["17"] != nil {
 		x.OtherTimer = new(durationpb.Duration)
-		err = ddb.UnmarshalMessage(m["17"], x.OtherTimer, ddb.Embed(v1.Encoding_ENCODING_UNSPECIFIED))
+		err = ddb.UnmarshalMessage(m["17"], x.OtherTimer, ddb.Embed(v1.Encoding_ENCODING_DYNAMO))
 		if err != nil {
 			return fmt.Errorf("failed to unmarshal field 'OtherTimer': %w", err)
 		}
